Sun, Jan 15th - Logout/ displayInfo()
- Created the logout feature which is just a button click. Event
listener for click on this button calls a logout function which
terminates the sessions and refreshes the page.

- On document.ready, session variables are defaulted to empty and
logged in = false if user if not already logged in, or no session
active.

- Based on which ‘$_SESSION[‘User’]’ is retireived, the page content is
displayed accordingly (as of right now for testing just have text
displayed , ie. “This is the Instructor page”, of which we would late
add content the instructor would see.