Mon, Jan 16th - page buttons/ replaced lost CSS
- added feature upon login, buttons appear which will link different
pages once clicked. These can be changed at anytime, can’t currently
remember the exact pages we agreed on.

- Being that the login table is generated and not placed in the
original main.php page, an event listener starts after the table is
placed to recognize when the submit button is pressed. Then sends the
information via Ajax. However this only allows the information to be
submitted via the button press, and hitting enter doesn’t not submit as
it is not a form.

- replaced some sample/ test CSS (very minor) that must have been
accidentally not saved properly or uploaded. eg. Separating top
elements of page in the banner evenly and entered across the entirety
of the page. Also changed size of UofTLogo