Jan 23rd - Display course/users,  delete course
-Developed 2 main content functions in Controller.js
(displayButtons(role) and displayPageInfo(page). displayButtons is
called only when the login_table form is submitted (via AJAX), as then
the page decides what tabs are to be displayed.

-Each of the buttons has their own onclick attribute which is dependent
on what page they are. For instance: the users button when logged in as
an ADMIN, when clicked calls displayPageInfo(“Users”). Note that we are
calling it with the single paramater “string”, as that tells
displayPageInfo(page), which page content we are specifically
interested in.

-From here, the function displayPageInfo(role) makes an Ajax call to
“Controller.php” from which are to receive the information to display.

-When asynchronously communicating with Controller.php, it connects to
our CSC492 database, specifically the users table and makes a query to
select all information from it. Controller.php then loop through each
row of the query result, then constructs a HTML string containing a
table with the appropriate content and titles.

-This string is then sent back to Controller.js from which it is
handled via the success attribute in the Ajax call. From here, the
response is simply placed into the pageInfo div via Jquery’s .html()
function. Note that this replaces any information that is currently
there. This is how the content updates when we switch pages (press
different page buttons).

-Problem: Jquery wasn’t able to find certain content because it wasn’t
made available until after Document.ready( ).
Solution: found out there was a .on(event,function) method that could
be called in Jquery.     It allowed us to be able to find and access
content on the DOM which wasn’t made
readily available on document.ready( ) which we initially had it as.
These queries no longer need to be based on time now which is simpler.

-Now added feature where instructors and Admins may remove courses. It
uses the primary key ‘CID’ in course table to be able to identify which
course should be removed. When the delete button is clicked, a function
deleteCourse( ) is called which genereates a new string and then calls
displayInfo ( ) which a few more variables sent to Contorller.php.
Here, controller.php check if the delete post variable set, and if so,
makes the delete query on the given CID from deleteCourse( ).