Jan 30th - Added modify Instructor feature.
- Added the modify instructor feature where ADMINS and teachers can now
selected a row represent which course they’d like to edit, and provided
a drop down list of all the teachers they may choose from of which
which change when they hit the change button.

- This calls the changeCourseIns() function which populates a new
string to send to displayPageInfo() (which is the Ajax function that
communicated with our controller.php.

- New Feature: added getCourseRowInfo() which reads the information of
each <td> in the selected row of the course table and stores each
element in global variables for reference from other functions. This is
a cleaner approach as it removes duplicate code whereas we were
generating and looking looping through this row every single time each
each alternate function.