Feb 14 - Application Process
Problem: Wasn’t able to use a session variable across multiple pages
(eg. $_SESSION[‘role’] from connect2db.php could not be used in
Controller.php as an index error was returned).
Solution: Had to add a “require_once” statement to be able to to use
the variables across both pages. This allows Controller.php to import
all Session variables from connect2db

Added 2 new relations.
1.Answers(AID, Num_Courses, Worked_With, TA_before, Mark_Received)
which stores the answers to quick questions in the form of the sent
applications. AID is just the reference primary key, Num_courses is how
many courses have been taken with the respective professor before,
Work_With is whether or not the applicant has worked for the professor
before, and TA_before is whether or not the applicant have Ta’d for
this course before.
2.Applications(UTORID,CID, Late, AID, Tag) which is essentially the
information of each application and applicant attached to it, just
separating it from the answers so this isn’t one large relation, and
instead 2 smaller, cleaner ones. The foreign keys are as follows:
Applications[UTORID]  Users[UTORID],  Applications[CID]  Course[CID],
Applications[AID]  Answers[AID].

When applicants select a course from the course list, when they click
apply, a form pops up giving a list of quick (mainly quantitative or
boolean questions to be stored for future reference by admins and
instructors). This will be separate from the profile of an applicant.

Added my Applications page which allows applicants to view all
applications they currently have (displays them on the page). Still
deciding if we would like to add a “status” feature for each
application. To see if granted, declined or pending etc. will ask Dan
in meeting. The results are assembled through a conditional natural
join of Answers,Applications and Course with the condition being they
must match the session user variable (only return this user’s
applications).