Feb 25 - Profile/ edit Profile/ force preferred choices
- Created the schema for a profile with a few temporary questions.
Profiles(PID, UTORID, YEAR_STUDY, TA_EXP, VOLUNTEER, BLURB). The
following foreign keys are Profiles[UTORID]  Users[UTORID], and PID
(profile Id) is the primary key. This permits only 1 profile per
applicant that may be updated and/or maintained.

- Noticed one neat feature about DBMS MARIAdb, the AUTO_INCREMENT
feature. This attribute characteristic is now placed on the column PID,
so now instead of looping to find the last highest iD number like is
done for the course table, once we input a new entry into the profiles
relation, it will automatically increment and create a new primary key
for us which saves us from having to make an additional query.

- Noticed since editing the course table to have the new attributes
keeping track of the number of jobs total and available, the add course
option is no longer working, so will need to eventually update that.

- Added My Profile page for applicant that allows users to view their
stored information. Note that this is also the same profile page that
is viewable by the instructors as well. It is not editable yet, but
just for testing, added profile information to the database via
terminal.

- Just like every other page function, the information (String) is sent
to displayPageInfo(page) in Controller.js which then communicates with
Controller.php for the appropriate information. A query is sent which
creates a natural join between Users and Profiles relations conditional
on the given UTORID to retrieve all information for the profile. Then
the retrieved data is constructed into HTML and sent back to
Controller.js to be spit onto the page for viewing only. A Button for
the editing option is also sent to edit the profile. Notice that this
button is only sent to the applicant, not the instructor as then the
instructor would also be given editing permissions which is not
necessary.

- Added feature: edit profile. Onlick of the “edit profile button”
gives you an editing option. Troubles: figuring a way to also keep the
current information in the profile so the user can see what they are
editing, as opposed to returning a bunch of empty text boxes and users
required to start from scratch. Solution: in controller.js ->
editProfile(), we have JQuery extract all the current values from the
profile, and then save them as variables for later use.

- Once the variables are saved, the it communicates to controller.php
for the edit profile form which is just blank, then after via JQuery,
assign the values of the corresponding text boxes the values that were
saved, so the applicant can see what they currently have saved. After,
when the user clicks on “Update Profile”, the information is sent back
to controller.php where an update statement is made to save the
information in the profile table.

- MAIN PROBLEM: the saved variables were not be added to the values of
the empty form for the user to view. After time realized once again,
this was because of the asynchronous AJAX call being made!!! So had to
make a callback procedure in success section of the displayPageInfo()
function essentially preventing the values from being place before the
empty profile edit page was finished being placed (whereas JQuery
wouldn’t find these tex tboxes being that there weren’t even placed
yet.) Now it works fine.

- The next tricky tasks, is adding, and forcing an applicant to pick up
to their top 5 application choices.

- Fixed the add courses, needed to add (position numbers) to the insert
statement and add those options into the form so their values can get
extracted

- Added a few error messages for instance when trying to add a user
without a utorid, will make the textbox red to signal the error
location and adds a message in the “errorMessage Div”, similar to
adding a user when the password and retype password doesn’t match.

- Add a few temporary colours in css, and positioned the tables created
in pageInfo div to have a wrapper which has a scroll feature… so an
abundance of data doesn’t continue to print on the page.. It will keep
it all on the single viewable page