Feb 28th - Application Status/ Granting Positions
- Add Application status feature. (Added select option as the last td
of every application row in the table #allAppTable). The value of this
select corresponds to the currently stored “Tag” value of the
application relation in the database. We have an event listener set up
which recognize when a prof/ admin goes to change one of these values
and sends the new information to Controller.Php

- In controller.php. A series of tests are conducted to ensure the
change is valid. Eg. if there are no positions available, it will
disallow you to grant another position to that course. Additionally if
the previous status was granted, and it is changed, that frees up a
position thus will update in the course table.

- The corresponding number of available positions for each course
(oppourunity) is updated instantly as application statues are changed.

Problem: when clicking on the select tag for an application status and
changing it, Jquery was unable to found our ‘selectedRow’ to extract
the necessary information for updating.
Solution: After conducting research, we came to the conclusion that
JQuery undergoes 	process called “event bubbling” which essentially
executes events in an inside out fashion. So we have 2 events that are
being called, onchange for the select and on click for the table row
(tr). Note that the select is located within the table row, thus it
will always get executed first. With this in mind, the onchange
callback function calls changeTags() before it even extracts the
information from the row (the tr listener). Thus to override this, we
changed the tr event to be mouse down and the select listener to be on
focus so that it extracts the row information first before it calls the
chageTags() function once the select has been changed.

- We also noticed now the sort function does not spit out the tag
options when it returns the sorted table so we will need to update that.

- ALSO, we didn’t realize we were using an outdated version of JQuery
so we have recently installed and included the most updated version.
(3.1.1.min.js)