Mar 10 - CSV complete/ Email
-BIG FEATURE: Admins can now go to the other page and upload a list of
courses via the other tab. Main problem which took several hours of
research, JQuery was not able to extract data from the uploaded .csv
file (for any browser). After hours of research, the simple reasoning
was that JQuery alone did not support this type of extraction. So the
solution was just to search for the data via the Document
(getElementById()).

- After, the upload was only working on Safari browser specifically. To
get around this, has to change the .on(‘submit’,’#input-file’) listener
to an .on(‘click’,’#input-file’) which fixed this problem.

- Added <pre> tag for profile views as we found out this tag preserves
indentation/ newlines and bullets etc. unlike <p> tag.

- Added email attribute to USERS table, which in addition added this
field to adding a user form, and is displayed in the user table.

- When an admin visits the profile of an applicant, they have a link to
which they may click and sends an email to that student. *May
eventually have our own email box straight on the site instead of it
opening a 3rd party mail app* (note students only see the text of their
email address on their profile, not the link to email.