March 10 - CSS, Tags, Integer logic
- Fixed tags displaying the wrong value if a user selected a new Status
and then cancelled it

- Problem: When accepting grade values during a course application
process, strings or invalid numbers could be input by the applicant.
Solution: We realized grade.val would still return a string so we used
parseInt to make it into a comparable number, then discovered that NaN
!= NaN and that you have to use isNaN as a comparison to prevent
problems with non-numerical inputs, and finally settled on
isNaN(parseInt(grade.val)), which works with every input we’ve tested.

- Multiple CSS improvements, reworked and added ids to many forms and
tables - still a work in progress.

- Implemented a maximum character length for answering profile
questions.