March 25th - Edit Applications/ Edit/view Course Questions
- As posted before, the admin and instructors are now able to see the
corresponding questions of a selected row. They are not able to edit
that values in the text boxes and submit to update their questions.

- Users when applying to a course they have already applied to, receive
a message informing them and asking if they wish to proceed and
overwrite their previous application information. The date is now also
updated as well.

- In the “my applications page” from the perspective of an applicant,
they may click on an application and choose to edit the information.
Similarly the data is also updated.

- when the number of courses and/or applications exceeds the amount of
space provided in the table on the screen, an overflow: scroll feature
is added to now the table will not extend off the page.